+++ sdk/core/azure-core-rest/src/main/java/com/azure/android/core/rest/implementation/ItemPage.java
@@ -3,7 +3,7 @@
 
 package com.azure.android.core.rest.implementation;
 
-import com.azure.android.core.rest.Page;
+import com.azure.android.core.util.paging.Page;
 import com.fasterxml.jackson.annotation.JsonAlias;
 
 import java.util.Collections;
